requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
idna==2.9
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.9
yagmail==0.11.224
